fix(#9): change the type of prioriy of basetime.

Closes #9

See merge request intdash-sdk/iscp-go-develop!22

+++ encoding/convert/proto_to_wire.go
@@ -677,7 +677,7 @@ func ToBaseTime(v *autogen.BaseTime) *message.BaseTime {
 	return &message.BaseTime{
 		SessionID:   v.SessionId,
 		Name:        v.Name,
-		Priority:    v.Priority,
+		Priority:    uint8(v.Priority),
 		ElapsedTime: time.Duration(v.ElapsedTime),
 		BaseTime:    time.Unix(0, v.BaseTime).UTC(),
 	}
